% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walrus.R
\name{:=}
\alias{:=}
\title{Walrus operator}
\usage{
`:=`(sym, val)
}
\arguments{
\item{sym}{The symbol to assign to}

\item{val}{The value to assign}
}
\value{
The assigned value
}
\description{
Walrus operator
}
\examples{
Letters := Enum(letters)
}
