% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum.R
\name{enum}
\alias{enum}
\alias{Enum}
\title{Enums}
\usage{
Enum(name, ...)

enum(values, ...)
}
\arguments{
\item{name}{The name of the `enum`}

\item{values, ...}{Set values for the `enum`.  These can be names or unnamed.}
}
\value{
An `enum` object, which is a function with S3 methods
}
\description{
Create enum sets
}
\details{
`enum`s are sets of named or unnamed values.  The `enum` object has built-in
features to retrieve values by name, provide a list of all values,
}
\examples{
Letters <- enum(letters)
Letters()        # when missing, returns list of everything
Letters(Letters) # pass the enum itself to retrieve just the values

# value, ... are combined:
Continents <- enum(
  "Africa",
  "Antarctica",
  "Asia",
  "Europe",
  "North America",
  "Oceania",
  "South America"
)

Continents()
Continents@Asia
Continents("Asia")
Continents(c("Asia", "Europe"))
try(Continents("Atlantis"))

Colors := Enum(c(
  Red = "#FF0000",
  Green = "#00FF00",
  Blue = "#0000FF",
  Yellow = "#FFFF00",
  Black = "#000000",
  White = "#FFFFFF",
  Purple = "#800080",
  Orange = "#FFA500",
  Pink = "#FFC0CB",
  Brown = "#A52A2A",
  Gray = "#808080",
  Cyan = "#00FFFF",
  Magenta = "#FF00FF"
))

Colors@Red
Colors(c("Red", "Blue"))
Colors("Green")
try(Colors("Turquoise"))  # This will raise an error

Shapes <- enum(c(
  Circle = "A round shape",
  Square = "A shape with four equal sides",
  Triangle = "A shape with three sides",
  Rectangle = "A shape with four sides and equal opposite sides",
  Oval = "An elongated circle",
  Star = "A shape with points radiating from a center",
  Heart = "A shape symbolizing love"
))

Shapes@Circle
Shapes(c("Circle", "Square"))
Colors(Colors)
print((function(x = Colors) Colors(x))())
print((function(x = Colors) Colors(x))("Red"))
}
