% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enum-class.R
\name{Enum_class}
\alias{Enum_class}
\title{Enum S7 Class}
\usage{
Enum_class(values = list(), enum_names = character(0))
}
\arguments{
\item{values}{A list of values.}

\item{enum_names}{A character vector of enum names.}
}
\value{
An \code{Enum} object.
}
\description{
An S7 class representing an enumeration with named values.
}
\examples{
colors <- Enum_class(values = list(RED = 1, GREEN = 2, BLUE = 3), enum_names = c("RED", "GREEN", "BLUE"))
}
