% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{is.valid}
\alias{is.valid}
\title{Check if a value is valid for an enum}
\usage{
is.valid(enum, value)
}
\arguments{
\item{enum}{An Enum object}

\item{value}{The value to check}
}
\value{
\code{TRUE} if the value is valid for the enum, \code{FALSE} otherwise
}
\description{
Check if a value is valid for an enum
}
\examples{
colors <- Enum(RED = 1, GREEN = 2, BLUE = 3)
is.valid(colors, 1) # TRUE
is.valid(colors, 4) # FALSE

}
