% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{key.by.value}
\alias{key.by.value}
\title{Get the key (name) for a given value in an enum}
\usage{
key.by.value(enum, value)
}
\arguments{
\item{enum}{An Enum object}

\item{value}{The value to find the key for}
}
\value{
The key (name) for the value, or \code{NULL} if not found
}
\description{
Get the key (name) for a given value in an enum
}
\examples{
colors <- Enum(RED = 1, GREEN = 2, BLUE = 3)
key.by.value(colors, 1) # "RED"
key.by.value(colors, 4) # NULL

}
