% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum-package.R
\docType{package}
\name{enum-package}
\alias{enum-package}
\title{enum: Enum Classes}
\description{
Creates enum classes. Enum classes are controls sets that allow users to specify specific choices for options or categories. Enums are present in many other programming languages and offer more support than using factors.
}
\author{
\strong{Maintainer}: Jordan Mark Barbone \email{jmbarbone@gmail.com} (\href{https://orcid.org/0000-0001-9788-3628}{ORCID}) [copyright holder]

}
\keyword{internal}
