% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum-package.R
\docType{package}
\name{enum-package}
\alias{enum}
\alias{enum-package}
\title{enum: Enums in R}
\description{
Provides a robust implementation of enumerations (enums) for R, enabling the creation of type-safe, validated sets of named constants. Enums help prevent errors by constraining variables to predefined values and offer better code readability and maintainability.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/milanmlft/enum}
  \item Report bugs at \url{https://github.com/milanmlft/enum/issues}
}

}
\author{
\strong{Maintainer}: Milan Malfait \email{m.malfait@ucl.ac.uk} (\href{https://orcid.org/0000-0001-9144-3701}{ORCID})

}
\keyword{internal}
